Configurable media type options
This change introduces a new configuration option which allows to force the import to use
the high-resolution original instead of a preview version.

By default, JPG and PNG are now using the preview generated by Pixxio instead of the original
file. This can improve import speed with very large photos and solve memory problems caused
by the resizing process.